// src/app/services/interfaces/interfaceRegister.ts
/**
 * Interface pour les champs d'inscription
 * Cette interface définit la structure des champs utilisés dans le processus d'inscription des utilisateurs.
 * Elle inclut des propriétés pour le label, le placeholder et le type du champ.
 */
export interface RegistrationField {
    label: string;
    placeholder: string;
    type: string;
}